refactor: 팁탭 리팩터링
